backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
openai
slowapi
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx
